refactor: split up retreival of data into seperate modules based on fact if team or player related
